Crate rusoto_ses

Amazon Simple Email Service
This document contains reference information for the Amazon Simple Email Service (Amazon SES) API, version 2010-12-01. This document is best used in conjunction with the Amazon SES Developer Guide.

For a list of Amazon SES endpoints to use in service requests, see Regions and Amazon SES in the Amazon SES Developer Guide.

If you're using the service, you're probably looking for SesClient and Ses.

Structs

AddHeaderAction
When included in a receipt rule, this action adds a header to the received email.

For information about adding a header using a receipt rule, see the Amazon SES Developer Guide.

Body
Represents the body of the message. You can specify text, HTML, or both. If you use both, then the message should display correctly in the widest variety of email clients.

BounceAction
When included in a receipt rule, this action rejects the received email by returning a bounce response to the sender and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

For information about sending a bounce message in response to a received email, see the Amazon SES Developer Guide.

BouncedRecipientInfo
Recipient-related information to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

BulkEmailDestination
An array that contains one or more Destinations, as well as the tags and replacement data associated with each of those Destinations.

BulkEmailDestinationStatus
An object that contains the response from the SendBulkTemplatedEmail operation.

CloneReceiptRuleSetRequest
Represents a request to create a receipt rule set by cloning an existing one. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

CloneReceiptRuleSetResponse
An empty element returned on a successful request.

CloudWatchDestination
Contains information associated with an Amazon CloudWatch event destination to which email sending events are published.

Event destinations, such as Amazon CloudWatch, are associated with configuration sets, which enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

CloudWatchDimensionConfiguration
Contains the dimension configuration to use when you publish email sending events to Amazon CloudWatch.

For information about publishing email sending events to Amazon CloudWatch, see the Amazon SES Developer Guide.

ConfigurationSet
The name of the configuration set.

Configuration sets let you create groups of rules that you can apply to the emails you send using Amazon SES. For more information about using configuration sets, see Using Amazon SES Configuration Sets in the Amazon SES Developer Guide.

Content
Represents textual data, plus an optional character set specification.

By default, the text must be 7-bit ASCII, due to the constraints of the SMTP protocol. If the text must contain any other characters, then you must also specify a character set. Examples include UTF-8, ISO-8859-1, and Shift_JIS.

CreateConfigurationSetEventDestinationRequest
Represents a request to create a configuration set event destination. A configuration set event destination, which can be either Amazon CloudWatch or Amazon Kinesis Firehose, describes an AWS service in which Amazon SES publishes the email sending events associated with a configuration set. For information about using configuration sets, see the Amazon SES Developer Guide.

CreateConfigurationSetEventDestinationResponse
An empty element returned on a successful request.

CreateConfigurationSetRequest
Represents a request to create a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

CreateConfigurationSetResponse
An empty element returned on a successful request.

CreateConfigurationSetTrackingOptionsRequest
Represents a request to create an open and click tracking option object in a configuration set.

CreateConfigurationSetTrackingOptionsResponse
An empty element returned on a successful request.

CreateCustomVerificationEmailTemplateRequest
Represents a request to create a custom verification email template.

CreateReceiptFilterRequest
Represents a request to create a new IP address filter. You use IP address filters when you receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

CreateReceiptFilterResponse
An empty element returned on a successful request.

CreateReceiptRuleRequest
Represents a request to create a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

CreateReceiptRuleResponse
An empty element returned on a successful request.

CreateReceiptRuleSetRequest
Represents a request to create an empty receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

CreateReceiptRuleSetResponse
An empty element returned on a successful request.

CreateTemplateRequest
Represents a request to create an email template. For more information, see the Amazon SES Developer Guide.

CreateTemplateResponse

CustomVerificationEmailTemplate
Contains information about a custom verification email template.

DeleteConfigurationSetEventDestinationRequest
Represents a request to delete a configuration set event destination. Configuration set event destinations are associated with configuration sets, which enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

DeleteConfigurationSetEventDestinationResponse
An empty element returned on a successful request.

DeleteConfigurationSetRequest
Represents a request to delete a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

DeleteConfigurationSetResponse
An empty element returned on a successful request.

DeleteConfigurationSetTrackingOptionsRequest
Represents a request to delete open and click tracking options in a configuration set.

DeleteConfigurationSetTrackingOptionsResponse
An empty element returned on a successful request.

DeleteCustomVerificationEmailTemplateRequest
Represents a request to delete an existing custom verification email template.

DeleteIdentityPolicyRequest
Represents a request to delete a sending authorization policy for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

DeleteIdentityPolicyResponse
An empty element returned on a successful request.

DeleteIdentityRequest
Represents a request to delete one of your Amazon SES identities (an email address or domain).

DeleteIdentityResponse
An empty element returned on a successful request.

DeleteReceiptFilterRequest
Represents a request to delete an IP address filter. You use IP address filters when you receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DeleteReceiptFilterResponse
An empty element returned on a successful request.

DeleteReceiptRuleRequest
Represents a request to delete a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DeleteReceiptRuleResponse
An empty element returned on a successful request.

DeleteReceiptRuleSetRequest
Represents a request to delete a receipt rule set and all of the receipt rules it contains. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DeleteReceiptRuleSetResponse
An empty element returned on a successful request.

DeleteTemplateRequest
Represents a request to delete an email template. For more information, see the Amazon SES Developer Guide.

DeleteTemplateResponse

DeleteVerifiedEmailAddressRequest
Represents a request to delete an email address from the list of email addresses you have attempted to verify under your AWS account.

DeliveryOptions
Specifies whether messages that use the configuration set are required to use Transport Layer Security (TLS).

DescribeActiveReceiptRuleSetRequest
Represents a request to return the metadata and receipt rules for the receipt rule set that is currently active. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DescribeActiveReceiptRuleSetResponse
Represents the metadata and receipt rules for the receipt rule set that is currently active.

DescribeConfigurationSetRequest
Represents a request to return the details of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

DescribeConfigurationSetResponse
Represents the details of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

DescribeReceiptRuleRequest
Represents a request to return the details of a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DescribeReceiptRuleResponse
Represents the details of a receipt rule.

DescribeReceiptRuleSetRequest
Represents a request to return the details of a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

DescribeReceiptRuleSetResponse
Represents the details of the specified receipt rule set.

Destination

Represents the destination of the message, consisting of To:, CC:, and BCC: fields.

Amazon SES does not support the SMTPUTF8 extension, as described in RFC6531. For this reason, the local part of a destination email address (the part of the email address that precedes the @ sign) may only contain 7-bit ASCII characters. If the domain part of an address (the part after the @ sign) contains non-ASCII characters, they must be encoded using Punycode, as described in RFC3492.

EventDestination
Contains information about the event destination that the specified email sending events will be published to.

When you create or update an event destination, you must provide one, and only one, destination. The destination can be Amazon CloudWatch, Amazon Kinesis Firehose or Amazon Simple Notification Service (Amazon SNS).

Event destinations are associated with configuration sets, which enable you to publish email sending events to Amazon CloudWatch, Amazon Kinesis Firehose, or Amazon Simple Notification Service (Amazon SNS). For information about using configuration sets, see the Amazon SES Developer Guide.

ExtensionField
Additional X-headers to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

GetAccountSendingEnabledResponse
Represents a request to return the email sending status for your Amazon SES account in the current AWS Region.

GetCustomVerificationEmailTemplateRequest
Represents a request to retrieve an existing custom verification email template.

GetCustomVerificationEmailTemplateResponse
The content of the custom verification email template.

GetIdentityDkimAttributesRequest
Represents a request for the status of Amazon SES Easy DKIM signing for an identity. For domain identities, this request also returns the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES successfully verified that these tokens were published. For more information about Easy DKIM, see the Amazon SES Developer Guide.

GetIdentityDkimAttributesResponse
Represents the status of Amazon SES Easy DKIM signing for an identity. For domain identities, this response also contains the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES successfully verified that these tokens were published.

GetIdentityMailFromDomainAttributesRequest
Represents a request to return the Amazon SES custom MAIL FROM attributes for a list of identities. For information about using a custom MAIL FROM domain, see the Amazon SES Developer Guide.

GetIdentityMailFromDomainAttributesResponse
Represents the custom MAIL FROM attributes for a list of identities.

GetIdentityNotificationAttributesRequest
Represents a request to return the notification attributes for a list of identities you verified with Amazon SES. For information about Amazon SES notifications, see the Amazon SES Developer Guide.

GetIdentityNotificationAttributesResponse
Represents the notification attributes for a list of identities.

GetIdentityPoliciesRequest
Represents a request to return the requested sending authorization policies for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

GetIdentityPoliciesResponse
Represents the requested sending authorization policies.

GetIdentityVerificationAttributesRequest
Represents a request to return the Amazon SES verification status of a list of identities. For domain identities, this request also returns the verification token. For information about verifying identities with Amazon SES, see the Amazon SES Developer Guide.

GetIdentityVerificationAttributesResponse
The Amazon SES verification status of a list of identities. For domain identities, this response also contains the verification token.

GetSendQuotaResponse
Represents your Amazon SES daily sending quota, maximum send rate, and the number of emails you have sent in the last 24 hours.

GetSendStatisticsResponse
Represents a list of data points. This list contains aggregated data from the previous two weeks of your sending activity with Amazon SES.

GetTemplateRequest

GetTemplateResponse

IdentityDkimAttributes
Represents the DKIM attributes of a verified email address or a domain.

IdentityMailFromDomainAttributes
Represents the custom MAIL FROM domain attributes of a verified identity (email address or domain).

IdentityNotificationAttributes
Represents the notification attributes of an identity, including whether an identity has Amazon Simple Notification Service (Amazon SNS) topics set for bounce, complaint, and/or delivery notifications, and whether feedback forwarding is enabled for bounce and complaint notifications.

IdentityVerificationAttributes
Represents the verification attributes of a single identity.

KinesisFirehoseDestination
Contains the delivery stream ARN and the IAM role ARN associated with an Amazon Kinesis Firehose event destination.

Event destinations, such as Amazon Kinesis Firehose, are associated with configuration sets, which enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

LambdaAction
When included in a receipt rule, this action calls an AWS Lambda function and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

To enable Amazon SES to call your AWS Lambda function or to publish to an Amazon SNS topic of another account, Amazon SES must have permission to access those resources. For information about giving permissions, see the Amazon SES Developer Guide.

For information about using AWS Lambda actions in receipt rules, see the Amazon SES Developer Guide.

ListConfigurationSetsRequest
Represents a request to list the configuration sets associated with your AWS account. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

ListConfigurationSetsResponse
A list of configuration sets associated with your AWS account. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

ListCustomVerificationEmailTemplatesRequest
Represents a request to list the existing custom verification email templates for your account.

For more information about custom verification email templates, see Using Custom Verification Email Templates in the Amazon SES Developer Guide.

ListCustomVerificationEmailTemplatesResponse
A paginated list of custom verification email templates.

ListIdentitiesRequest
Represents a request to return a list of all identities (email addresses and domains) that you have attempted to verify under your AWS account, regardless of verification status.

ListIdentitiesResponse
A list of all identities that you have attempted to verify under your AWS account, regardless of verification status.

ListIdentityPoliciesRequest
Represents a request to return a list of sending authorization policies that are attached to an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

ListIdentityPoliciesResponse
A list of names of sending authorization policies that apply to an identity.

ListReceiptFiltersRequest
Represents a request to list the IP address filters that exist under your AWS account. You use IP address filters when you receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

ListReceiptFiltersResponse
A list of IP address filters that exist under your AWS account.

ListReceiptRuleSetsRequest
Represents a request to list the receipt rule sets that exist under your AWS account. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

ListReceiptRuleSetsResponse
A list of receipt rule sets that exist under your AWS account.

ListTemplatesRequest

ListTemplatesResponse

ListVerifiedEmailAddressesResponse
A list of email addresses that you have verified with Amazon SES under your AWS account.

Message
Represents the message to be sent, composed of a subject and a body.

MessageDsn
Message-related information to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

MessageTag
Contains the name and value of a tag that you can provide to SendEmail or SendRawEmail to apply to an email.

Message tags, which you use with configuration sets, enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

PutConfigurationSetDeliveryOptionsRequest
A request to modify the delivery options for a configuration set.

PutConfigurationSetDeliveryOptionsResponse
An HTTP 200 response if the request succeeds, or an error message if the request fails.

PutIdentityPolicyRequest
Represents a request to add or update a sending authorization policy for an identity. Sending authorization is an Amazon SES feature that enables you to authorize other senders to use your identities. For information, see the Amazon SES Developer Guide.

PutIdentityPolicyResponse
An empty element returned on a successful request.

RawMessage
Represents the raw data of the message.

ReceiptAction
An action that Amazon SES can take when it receives an email on behalf of one or more email addresses or domains that you own. An instance of this data type can represent only one action.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

ReceiptFilter
A receipt IP address filter enables you to specify whether to accept or reject mail originating from an IP address or range of IP addresses.

For information about setting up IP address filters, see the Amazon SES Developer Guide.

ReceiptIpFilter
A receipt IP address filter enables you to specify whether to accept or reject mail originating from an IP address or range of IP addresses.

For information about setting up IP address filters, see the Amazon SES Developer Guide.

ReceiptRule
Receipt rules enable you to specify which actions Amazon SES should take when it receives mail on behalf of one or more email addresses or domains that you own.

Each receipt rule defines a set of email addresses or domains that it applies to. If the email addresses or domains match at least one recipient address of the message, Amazon SES executes all of the receipt rule's actions on the message.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

ReceiptRuleSetMetadata
Information about a receipt rule set.

A receipt rule set is a collection of rules that specify what Amazon SES should do with mail it receives on behalf of your account's verified domains.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide.

RecipientDsnFields
Recipient-related information to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

ReorderReceiptRuleSetRequest
Represents a request to reorder the receipt rules within a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

ReorderReceiptRuleSetResponse
An empty element returned on a successful request.

ReputationOptions
Contains information about the reputation settings for a configuration set.

S3Action
When included in a receipt rule, this action saves the received message to an Amazon Simple Storage Service (Amazon S3) bucket and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

To enable Amazon SES to write emails to your Amazon S3 bucket, use an AWS KMS key to encrypt your emails, or publish to an Amazon SNS topic of another account, Amazon SES must have permission to access those resources. For information about giving permissions, see the Amazon SES Developer Guide.

When you save your emails to an Amazon S3 bucket, the maximum email size (including headers) is 30 MB. Emails larger than that will bounce.

For information about specifying Amazon S3 actions in receipt rules, see the Amazon SES Developer Guide.

SNSAction
When included in a receipt rule, this action publishes a notification to Amazon Simple Notification Service (Amazon SNS). This action includes a complete copy of the email content in the Amazon SNS notifications. Amazon SNS notifications for all other actions simply provide information about the email. They do not include the email content itself.

If you own the Amazon SNS topic, you don't need to do anything to give Amazon SES permission to publish emails to it. However, if you don't own the Amazon SNS topic, you need to attach a policy to the topic to give Amazon SES permissions to access it. For information about giving permissions, see the Amazon SES Developer Guide.

You can only publish emails that are 150 KB or less (including the header) to Amazon SNS. Larger emails will bounce. If you anticipate emails larger than 150 KB, use the S3 action instead.

For information about using a receipt rule to publish an Amazon SNS notification, see the Amazon SES Developer Guide.

SNSDestination
Contains the topic ARN associated with an Amazon Simple Notification Service (Amazon SNS) event destination.

Event destinations, such as Amazon SNS, are associated with configuration sets, which enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

SendBounceRequest
Represents a request to send a bounce message to the sender of an email you received through Amazon SES.

SendBounceResponse
Represents a unique message ID.

SendBulkTemplatedEmailRequest
Represents a request to send a templated email to multiple destinations using Amazon SES. For more information, see the Amazon SES Developer Guide.

SendBulkTemplatedEmailResponse

SendCustomVerificationEmailRequest
Represents a request to send a custom verification email to a specified recipient.

SendCustomVerificationEmailResponse
The response received when attempting to send the custom verification email.

SendDataPoint
Represents sending statistics data. Each SendDataPoint contains statistics for a 15-minute period of sending activity.

SendEmailRequest
Represents a request to send a single formatted email using Amazon SES. For more information, see the Amazon SES Developer Guide.

SendEmailResponse
Represents a unique message ID.

SendRawEmailRequest
Represents a request to send a single raw email using Amazon SES. For more information, see the Amazon SES Developer Guide.

SendRawEmailResponse
Represents a unique message ID.

SendTemplatedEmailRequest
Represents a request to send a templated email using Amazon SES. For more information, see the Amazon SES Developer Guide.

SendTemplatedEmailResponse

SesClient
A client for the Amazon SES API.

SetActiveReceiptRuleSetRequest
Represents a request to set a receipt rule set as the active receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

SetActiveReceiptRuleSetResponse
An empty element returned on a successful request.

SetIdentityDkimEnabledRequest
Represents a request to enable or disable Amazon SES Easy DKIM signing for an identity. For more information about setting up Easy DKIM, see the Amazon SES Developer Guide.

SetIdentityDkimEnabledResponse
An empty element returned on a successful request.

SetIdentityFeedbackForwardingEnabledRequest
Represents a request to enable or disable whether Amazon SES forwards you bounce and complaint notifications through email. For information about email feedback forwarding, see the Amazon SES Developer Guide.

SetIdentityFeedbackForwardingEnabledResponse
An empty element returned on a successful request.

SetIdentityHeadersInNotificationsEnabledRequest
Represents a request to set whether Amazon SES includes the original email headers in the Amazon SNS notifications of a specified type. For information about notifications, see the Amazon SES Developer Guide.

SetIdentityHeadersInNotificationsEnabledResponse
An empty element returned on a successful request.

SetIdentityMailFromDomainRequest
Represents a request to enable or disable the Amazon SES custom MAIL FROM domain setup for a verified identity. For information about using a custom MAIL FROM domain, see the Amazon SES Developer Guide.

SetIdentityMailFromDomainResponse
An empty element returned on a successful request.

SetIdentityNotificationTopicRequest
Represents a request to specify the Amazon SNS topic to which Amazon SES will publish bounce, complaint, or delivery notifications for emails sent with that identity as the Source. For information about Amazon SES notifications, see the Amazon SES Developer Guide.

SetIdentityNotificationTopicResponse
An empty element returned on a successful request.

SetReceiptRulePositionRequest
Represents a request to set the position of a receipt rule in a receipt rule set. You use receipt rule sets to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

SetReceiptRulePositionResponse
An empty element returned on a successful request.

StopAction
When included in a receipt rule, this action terminates the evaluation of the receipt rule set and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

For information about setting a stop action in a receipt rule, see the Amazon SES Developer Guide.

Template
The content of the email, composed of a subject line, an HTML part, and a text-only part.

TemplateMetadata
Contains information about an email template.

TestRenderTemplateRequest

TestRenderTemplateResponse

TrackingOptions
A domain that is used to redirect email recipients to an Amazon SES-operated domain. This domain captures open and click events generated by Amazon SES emails.

For more information, see Configuring Custom Domains to Handle Open and Click Tracking in the Amazon SES Developer Guide.

UpdateAccountSendingEnabledRequest
Represents a request to enable or disable the email sending capabilities for your entire Amazon SES account.

UpdateConfigurationSetEventDestinationRequest
Represents a request to update the event destination of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

UpdateConfigurationSetEventDestinationResponse
An empty element returned on a successful request.

UpdateConfigurationSetReputationMetricsEnabledRequest
Represents a request to modify the reputation metric publishing settings for a configuration set.

UpdateConfigurationSetSendingEnabledRequest
Represents a request to enable or disable the email sending capabilities for a specific configuration set.

UpdateConfigurationSetTrackingOptionsRequest
Represents a request to update the tracking options for a configuration set.

UpdateConfigurationSetTrackingOptionsResponse
An empty element returned on a successful request.

UpdateCustomVerificationEmailTemplateRequest
Represents a request to update an existing custom verification email template.

UpdateReceiptRuleRequest
Represents a request to update a receipt rule. You use receipt rules to receive email with Amazon SES. For more information, see the Amazon SES Developer Guide.

UpdateReceiptRuleResponse
An empty element returned on a successful request.

UpdateTemplateRequest

UpdateTemplateResponse

VerifyDomainDkimRequest
Represents a request to generate the CNAME records needed to set up Easy DKIM with Amazon SES. For more information about setting up Easy DKIM, see the Amazon SES Developer Guide.

VerifyDomainDkimResponse
Returns CNAME records that you must publish to the DNS server of your domain to set up Easy DKIM with Amazon SES.

VerifyDomainIdentityRequest
Represents a request to begin Amazon SES domain verification and to generate the TXT records that you must publish to the DNS server of your domain to complete the verification. For information about domain verification, see the Amazon SES Developer Guide.

VerifyDomainIdentityResponse
Returns a TXT record that you must publish to the DNS server of your domain to complete domain verification with Amazon SES.

VerifyEmailAddressRequest
Represents a request to begin email address verification with Amazon SES. For information about email address verification, see the Amazon SES Developer Guide.

VerifyEmailIdentityRequest
Represents a request to begin email address verification with Amazon SES. For information about email address verification, see the Amazon SES Developer Guide.

VerifyEmailIdentityResponse
An empty element returned on a successful request.

WorkmailAction
When included in a receipt rule, this action calls Amazon WorkMail and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS). You will typically not use this action directly because Amazon WorkMail adds the rule automatically during its setup procedure.

For information using a receipt rule to call Amazon WorkMail, see the Amazon SES Developer Guide.

Enums

CloneReceiptRuleSetError
Errors returned by CloneReceiptRuleSet

CreateConfigurationSetError
Errors returned by CreateConfigurationSet

CreateConfigurationSetEventDestinationError
Errors returned by CreateConfigurationSetEventDestination

CreateConfigurationSetTrackingOptionsError
Errors returned by CreateConfigurationSetTrackingOptions

CreateCustomVerificationEmailTemplateError
Errors returned by CreateCustomVerificationEmailTemplate

CreateReceiptFilterError
Errors returned by CreateReceiptFilter

CreateReceiptRuleError
Errors returned by CreateReceiptRule

CreateReceiptRuleSetError
Errors returned by CreateReceiptRuleSet

CreateTemplateError
Errors returned by CreateTemplate

DeleteConfigurationSetError
Errors returned by DeleteConfigurationSet

DeleteConfigurationSetEventDestinationError
Errors returned by DeleteConfigurationSetEventDestination

DeleteConfigurationSetTrackingOptionsError
Errors returned by DeleteConfigurationSetTrackingOptions

DeleteCustomVerificationEmailTemplateError
Errors returned by DeleteCustomVerificationEmailTemplate

DeleteIdentityError
Errors returned by DeleteIdentity

DeleteIdentityPolicyError
Errors returned by DeleteIdentityPolicy

DeleteReceiptFilterError
Errors returned by DeleteReceiptFilter

DeleteReceiptRuleError
Errors returned by DeleteReceiptRule

DeleteReceiptRuleSetError
Errors returned by DeleteReceiptRuleSet

DeleteTemplateError
Errors returned by DeleteTemplate

DeleteVerifiedEmailAddressError
Errors returned by DeleteVerifiedEmailAddress

DescribeActiveReceiptRuleSetError
Errors returned by DescribeActiveReceiptRuleSet

DescribeConfigurationSetError
Errors returned by DescribeConfigurationSet

DescribeReceiptRuleError
Errors returned by DescribeReceiptRule

DescribeReceiptRuleSetError
Errors returned by DescribeReceiptRuleSet

GetAccountSendingEnabledError
Errors returned by GetAccountSendingEnabled

GetCustomVerificationEmailTemplateError
Errors returned by GetCustomVerificationEmailTemplate

GetIdentityDkimAttributesError
Errors returned by GetIdentityDkimAttributes

GetIdentityMailFromDomainAttributesError
Errors returned by GetIdentityMailFromDomainAttributes

GetIdentityNotificationAttributesError
Errors returned by GetIdentityNotificationAttributes

GetIdentityPoliciesError
Errors returned by GetIdentityPolicies

GetIdentityVerificationAttributesError
Errors returned by GetIdentityVerificationAttributes

GetSendQuotaError
Errors returned by GetSendQuota

GetSendStatisticsError
Errors returned by GetSendStatistics

GetTemplateError
Errors returned by GetTemplate

ListConfigurationSetsError
Errors returned by ListConfigurationSets

ListCustomVerificationEmailTemplatesError
Errors returned by ListCustomVerificationEmailTemplates

ListIdentitiesError
Errors returned by ListIdentities

ListIdentityPoliciesError
Errors returned by ListIdentityPolicies

ListReceiptFiltersError
Errors returned by ListReceiptFilters

ListReceiptRuleSetsError
Errors returned by ListReceiptRuleSets

ListTemplatesError
Errors returned by ListTemplates

ListVerifiedEmailAddressesError
Errors returned by ListVerifiedEmailAddresses

PutConfigurationSetDeliveryOptionsError
Errors returned by PutConfigurationSetDeliveryOptions

PutIdentityPolicyError
Errors returned by PutIdentityPolicy

ReorderReceiptRuleSetError
Errors returned by ReorderReceiptRuleSet

SendBounceError
Errors returned by SendBounce

SendBulkTemplatedEmailError
Errors returned by SendBulkTemplatedEmail

SendCustomVerificationEmailError
Errors returned by SendCustomVerificationEmail

SendEmailError
Errors returned by SendEmail

SendRawEmailError
Errors returned by SendRawEmail

SendTemplatedEmailError
Errors returned by SendTemplatedEmail

SetActiveReceiptRuleSetError
Errors returned by SetActiveReceiptRuleSet

SetIdentityDkimEnabledError
Errors returned by SetIdentityDkimEnabled

SetIdentityFeedbackForwardingEnabledError
Errors returned by SetIdentityFeedbackForwardingEnabled

SetIdentityHeadersInNotificationsEnabledError
Errors returned by SetIdentityHeadersInNotificationsEnabled

SetIdentityMailFromDomainError
Errors returned by SetIdentityMailFromDomain

SetIdentityNotificationTopicError
Errors returned by SetIdentityNotificationTopic

SetReceiptRulePositionError
Errors returned by SetReceiptRulePosition

TestRenderTemplateError
Errors returned by TestRenderTemplate

UpdateAccountSendingEnabledError
Errors returned by UpdateAccountSendingEnabled

UpdateConfigurationSetEventDestinationError
Errors returned by UpdateConfigurationSetEventDestination

UpdateConfigurationSetReputationMetricsEnabledError
Errors returned by UpdateConfigurationSetReputationMetricsEnabled

UpdateConfigurationSetSendingEnabledError
Errors returned by UpdateConfigurationSetSendingEnabled

UpdateConfigurationSetTrackingOptionsError
Errors returned by UpdateConfigurationSetTrackingOptions

UpdateCustomVerificationEmailTemplateError
Errors returned by UpdateCustomVerificationEmailTemplate

UpdateReceiptRuleError
Errors returned by UpdateReceiptRule

UpdateTemplateError
Errors returned by UpdateTemplate

VerifyDomainDkimError
Errors returned by VerifyDomainDkim

VerifyDomainIdentityError
Errors returned by VerifyDomainIdentity

VerifyEmailAddressError
Errors returned by VerifyEmailAddress

VerifyEmailIdentityError
Errors returned by VerifyEmailIdentity

Traits

Ses
Trait representing the capabilities of the Amazon SES API. Amazon SES clients implement this trait.