Enum rusoto_ses::CreateReceiptRuleSetError

pub enum CreateReceiptRuleSetError {
    AlreadyExists(String),
    LimitExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateReceiptRuleSet

Variants
AlreadyExists(String)

Indicates that a resource could not be created because of a naming conflict.

LimitExceeded(String)

Indicates that a resource could not be created because of service limits. For a list of Amazon SES limits, see the Amazon SES Developer Guide.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateReceiptRuleSetError

pub fn from_body(body: &str) -> CreateReceiptRuleSetError