Struct rusoto_ses::CreateCustomVerificationEmailTemplateRequest

pub struct CreateCustomVerificationEmailTemplateRequest {
    pub failure_redirection_url: String,
    pub from_email_address: String,
    pub success_redirection_url: String,
    pub template_content: String,
    pub template_name: String,
    pub template_subject: String,
}

Represents a request to create a custom verification email template.

Fields
failure_redirection_url: String

The URL that the recipient of the verification email is sent to if his or her address is not successfully verified.

from_email_address: String

The email address that the custom verification email is sent from.

success_redirection_url: String

The URL that the recipient of the verification email is sent to if his or her address is successfully verified.

template_content: String

The content of the custom verification email. The total size of the email must be less than 10 MB. The message body may contain HTML, with some limitations. For more information, see Custom Verification Email Frequently Asked Questions in the Amazon SES Developer Guide.

template_name: String

The name of the custom verification email template.

template_subject: String

The subject line of the custom verification email.