Struct rusoto_ses::MessageDsn

pub struct MessageDsn {
    pub arrival_date: Option<String>,
    pub extension_fields: Option<Vec<ExtensionField>>,
    pub reporting_mta: String,
}

Message-related information to include in the Delivery Status Notification (DSN) when an email that Amazon SES receives on your behalf bounces.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

Fields
arrival_date: Option<String>

When the message was received by the reporting mail transfer agent (MTA), in RFC 822 date-time format.

extension_fields: Option<Vec<ExtensionField>>

Additional X-headers to include in the DSN.

reporting_mta: String

The reporting MTA that attempted to deliver the message, formatted as specified in RFC 3464 (mta-name-type; mta-name). The default value is dns; inbound-smtp.[region].amazonaws.com.