Struct rusoto_ses::ReceiptAction

pub struct ReceiptAction {
    pub add_header_action: Option<AddHeaderAction>,
    pub bounce_action: Option<BounceAction>,
    pub lambda_action: Option<LambdaAction>,
    pub s3_action: Option<S3Action>,
    pub sns_action: Option<SNSAction>,
    pub stop_action: Option<StopAction>,
    pub workmail_action: Option<WorkmailAction>,
}

An action that Amazon SES can take when it receives an email on behalf of one or more email addresses or domains that you own. An instance of this data type can represent only one action.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

Fields
add_header_action: Option<AddHeaderAction>

Adds a header to the received email.

bounce_action: Option<BounceAction>

Rejects the received email by returning a bounce response to the sender and, optionally, publishes a notification to Amazon Simple Notification Service (Amazon SNS).

lambda_action: Option<LambdaAction>

Calls an AWS Lambda function, and optionally, publishes a notification to Amazon SNS.

s3_action: Option<S3Action>

Saves the received message to an Amazon Simple Storage Service (Amazon S3) bucket and, optionally, publishes a notification to Amazon SNS.

sns_action: Option<SNSAction>

Publishes the email content within a notification to Amazon SNS.

stop_action: Option<StopAction>

Terminates the evaluation of the receipt rule set and optionally publishes a notification to Amazon SNS.

workmail_action: Option<WorkmailAction>

Calls Amazon WorkMail and, optionally, publishes a notification to Amazon SNS.