Struct rusoto_ses::SendEmailRequest

pub struct SendEmailRequest {
    pub configuration_set_name: Option<String>,
    pub destination: Destination,
    pub message: Message,
    pub reply_to_addresses: Option<Vec<String>>,
    pub return_path: Option<String>,
    pub return_path_arn: Option<String>,
    pub source: String,
    pub source_arn: Option<String>,
    pub tags: Option<Vec<MessageTag>>,
}

Represents a request to send a single formatted email using Amazon SES. For more information, see the Amazon SES Developer Guide.

Fields
configuration_set_name: Option<String>

The name of the configuration set to use when you send an email using SendEmail.

destination: Destination

The destination for this email, composed of To:, CC:, and BCC: fields.

message: Message

The message to be sent.

reply_to_addresses: Option<Vec<String>>

The reply-to email address(es) for the message. If the recipient replies to the message, each reply-to address will receive the reply.

return_path: Option<String>

The email address to which bounces and complaints are to be forwarded when feedback forwarding is enabled. If the message cannot be delivered to the recipient, then an error message will be returned from the recipient's ISP; this message will then be forwarded to the email address specified by the ReturnPath parameter. The ReturnPath parameter is never overwritten. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES.

return_path_arn: Option<String>

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

source: String

The email address that is sending the email. This email address must be either individually verified with Amazon SES, or from a domain that has been verified with Amazon SES. For information about verifying identities, see the Amazon SES Developer Guide.

If you are sending on behalf of another user and have been permitted to do so by a sending authorization policy, then you must also specify the SourceArn parameter. For more information about sending authorization, see the Amazon SES Developer Guide.

In all cases, the email address must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=. For more information, see RFC 2047.

source_arn: Option<String>

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

For more information about sending authorization, see the Amazon SES Developer Guide.

tags: Option<Vec<MessageTag>>

A list of tags, in the form of name/value pairs, to apply to an email that you send using SendEmail. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.