Struct rusoto_ses::SendRawEmailRequest

pub struct SendRawEmailRequest {
    pub configuration_set_name: Option<String>,
    pub destinations: Option<Vec<String>>,
    pub from_arn: Option<String>,
    pub raw_message: RawMessage,
    pub return_path_arn: Option<String>,
    pub source: Option<String>,
    pub source_arn: Option<String>,
    pub tags: Option<Vec<MessageTag>>,
}

Represents a request to send a single raw email using Amazon SES. For more information, see the Amazon SES Developer Guide.

Fields
configuration_set_name: Option<String>

The name of the configuration set to use when you send an email using SendRawEmail.

destinations: Option<Vec<String>>

A list of destinations for the message, consisting of To:, CC:, and BCC: addresses.

from_arn: Option<String>

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to specify a particular "From" address in the header of the raw email.

Instead of using this parameter, you can use the X-header X-SES-FROM-ARN in the raw message of the email. If you use both the FromArn parameter and the corresponding X-header, Amazon SES uses the value of the FromArn parameter.

For information about when to use this parameter, see the description of SendRawEmail in this guide, or see the Amazon SES Developer Guide.

raw_message: RawMessage

The raw text of the message. The client is responsible for ensuring the following:

Message must contain a header and a body, separated by a blank line.

All required header fields must be present.

Each part of a multipart MIME message must be formatted properly.

MIME content types must be among those supported by Amazon SES. For more information, go to the Amazon SES Developer Guide.

Must be base64-encoded.

return_path_arn: Option<String>

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to use the email address specified in the ReturnPath parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to use feedback@example.com, then you would specify the ReturnPathArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the ReturnPath to be feedback@example.com.

Instead of using this parameter, you can use the X-header X-SES-RETURN-PATH-ARN in the raw message of the email. If you use both the ReturnPathArn parameter and the corresponding X-header, Amazon SES uses the value of the ReturnPathArn parameter.

For information about when to use this parameter, see the description of SendRawEmail in this guide, or see the Amazon SES Developer Guide.

source: Option<String>

The identity's email address. If you do not provide a value for this parameter, you must specify a "From" address in the raw text of the message. (You can also specify both.)

By default, the string must be 7-bit ASCII. If the text must contain any other characters, then you must use MIME encoded-word syntax (RFC 2047) instead of a literal string. MIME encoded-word syntax uses the following form: =?charset?encoding?encoded-text?=. For more information, see RFC 2047.

If you specify the Source parameter and have feedback forwarding enabled, then bounces and complaints will be sent to this email address. This takes precedence over any Return-Path header that you might include in the raw text of the message.

source_arn: Option<String>

This parameter is used only for sending authorization. It is the ARN of the identity that is associated with the sending authorization policy that permits you to send for the email address specified in the Source parameter.

For example, if the owner of example.com (which has ARN arn:aws:ses:us-east-1:123456789012:identity/example.com) attaches a policy to it that authorizes you to send from user@example.com, then you would specify the SourceArn to be arn:aws:ses:us-east-1:123456789012:identity/example.com, and the Source to be user@example.com.

Instead of using this parameter, you can use the X-header X-SES-SOURCE-ARN in the raw message of the email. If you use both the SourceArn parameter and the corresponding X-header, Amazon SES uses the value of the SourceArn parameter.

For information about when to use this parameter, see the description of SendRawEmail in this guide, or see the Amazon SES Developer Guide.

tags: Option<Vec<MessageTag>>

A list of tags, in the form of name/value pairs, to apply to an email that you send using SendRawEmail. Tags correspond to characteristics of the email that you define, so that you can publish email sending events.