Struct rusoto_ses::UpdateConfigurationSetEventDestinationRequest

pub struct UpdateConfigurationSetEventDestinationRequest {
    pub configuration_set_name: String,
    pub event_destination: EventDestination,
}

Represents a request to update the event destination of a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

Fields
configuration_set_name: String

The name of the configuration set that you want to update.

event_destination: EventDestination

The event destination object that you want to apply to the specified configuration set.