Trait rusoto_ses::Ses

pub trait Ses {
    fn clone_receipt_rule_set(
        &self, 
        input: &CloneReceiptRuleSetRequest
    ) -> Result<CloneReceiptRuleSetResponse, CloneReceiptRuleSetError>;
    fn create_configuration_set(
        &self, 
        input: &CreateConfigurationSetRequest
    ) -> Result<CreateConfigurationSetResponse, CreateConfigurationSetError>;
    fn create_configuration_set_event_destination(
        &self, 
        input: &CreateConfigurationSetEventDestinationRequest
    ) -> Result<CreateConfigurationSetEventDestinationResponse, CreateConfigurationSetEventDestinationError>;
    fn create_receipt_filter(
        &self, 
        input: &CreateReceiptFilterRequest
    ) -> Result<CreateReceiptFilterResponse, CreateReceiptFilterError>;
    fn create_receipt_rule(
        &self, 
        input: &CreateReceiptRuleRequest
    ) -> Result<CreateReceiptRuleResponse, CreateReceiptRuleError>;
    fn create_receipt_rule_set(
        &self, 
        input: &CreateReceiptRuleSetRequest
    ) -> Result<CreateReceiptRuleSetResponse, CreateReceiptRuleSetError>;
    fn delete_configuration_set(
        &self, 
        input: &DeleteConfigurationSetRequest
    ) -> Result<DeleteConfigurationSetResponse, DeleteConfigurationSetError>;
    fn delete_configuration_set_event_destination(
        &self, 
        input: &DeleteConfigurationSetEventDestinationRequest
    ) -> Result<DeleteConfigurationSetEventDestinationResponse, DeleteConfigurationSetEventDestinationError>;
    fn delete_identity(
        &self, 
        input: &DeleteIdentityRequest
    ) -> Result<DeleteIdentityResponse, DeleteIdentityError>;
    fn delete_identity_policy(
        &self, 
        input: &DeleteIdentityPolicyRequest
    ) -> Result<DeleteIdentityPolicyResponse, DeleteIdentityPolicyError>;
    fn delete_receipt_filter(
        &self, 
        input: &DeleteReceiptFilterRequest
    ) -> Result<DeleteReceiptFilterResponse, DeleteReceiptFilterError>;
    fn delete_receipt_rule(
        &self, 
        input: &DeleteReceiptRuleRequest
    ) -> Result<DeleteReceiptRuleResponse, DeleteReceiptRuleError>;
    fn delete_receipt_rule_set(
        &self, 
        input: &DeleteReceiptRuleSetRequest
    ) -> Result<DeleteReceiptRuleSetResponse, DeleteReceiptRuleSetError>;
    fn delete_verified_email_address(
        &self, 
        input: &DeleteVerifiedEmailAddressRequest
    ) -> Result<(), DeleteVerifiedEmailAddressError>;
    fn describe_active_receipt_rule_set(
        &self, 
        input: &DescribeActiveReceiptRuleSetRequest
    ) -> Result<DescribeActiveReceiptRuleSetResponse, DescribeActiveReceiptRuleSetError>;
    fn describe_configuration_set(
        &self, 
        input: &DescribeConfigurationSetRequest
    ) -> Result<DescribeConfigurationSetResponse, DescribeConfigurationSetError>;
    fn describe_receipt_rule(
        &self, 
        input: &DescribeReceiptRuleRequest
    ) -> Result<DescribeReceiptRuleResponse, DescribeReceiptRuleError>;
    fn describe_receipt_rule_set(
        &self, 
        input: &DescribeReceiptRuleSetRequest
    ) -> Result<DescribeReceiptRuleSetResponse, DescribeReceiptRuleSetError>;
    fn get_identity_dkim_attributes(
        &self, 
        input: &GetIdentityDkimAttributesRequest
    ) -> Result<GetIdentityDkimAttributesResponse, GetIdentityDkimAttributesError>;
    fn get_identity_mail_from_domain_attributes(
        &self, 
        input: &GetIdentityMailFromDomainAttributesRequest
    ) -> Result<GetIdentityMailFromDomainAttributesResponse, GetIdentityMailFromDomainAttributesError>;
    fn get_identity_notification_attributes(
        &self, 
        input: &GetIdentityNotificationAttributesRequest
    ) -> Result<GetIdentityNotificationAttributesResponse, GetIdentityNotificationAttributesError>;
    fn get_identity_policies(
        &self, 
        input: &GetIdentityPoliciesRequest
    ) -> Result<GetIdentityPoliciesResponse, GetIdentityPoliciesError>;
    fn get_identity_verification_attributes(
        &self, 
        input: &GetIdentityVerificationAttributesRequest
    ) -> Result<GetIdentityVerificationAttributesResponse, GetIdentityVerificationAttributesError>;
    fn get_send_quota(&self) -> Result<GetSendQuotaResponse, GetSendQuotaError>;
    fn get_send_statistics(
        &self
    ) -> Result<GetSendStatisticsResponse, GetSendStatisticsError>;
    fn list_configuration_sets(
        &self, 
        input: &ListConfigurationSetsRequest
    ) -> Result<ListConfigurationSetsResponse, ListConfigurationSetsError>;
    fn list_identities(
        &self, 
        input: &ListIdentitiesRequest
    ) -> Result<ListIdentitiesResponse, ListIdentitiesError>;
    fn list_identity_policies(
        &self, 
        input: &ListIdentityPoliciesRequest
    ) -> Result<ListIdentityPoliciesResponse, ListIdentityPoliciesError>;
    fn list_receipt_filters(
        &self, 
        input: &ListReceiptFiltersRequest
    ) -> Result<ListReceiptFiltersResponse, ListReceiptFiltersError>;
    fn list_receipt_rule_sets(
        &self, 
        input: &ListReceiptRuleSetsRequest
    ) -> Result<ListReceiptRuleSetsResponse, ListReceiptRuleSetsError>;
    fn list_verified_email_addresses(
        &self
    ) -> Result<ListVerifiedEmailAddressesResponse, ListVerifiedEmailAddressesError>;
    fn put_identity_policy(
        &self, 
        input: &PutIdentityPolicyRequest
    ) -> Result<PutIdentityPolicyResponse, PutIdentityPolicyError>;
    fn reorder_receipt_rule_set(
        &self, 
        input: &ReorderReceiptRuleSetRequest
    ) -> Result<ReorderReceiptRuleSetResponse, ReorderReceiptRuleSetError>;
    fn send_bounce(
        &self, 
        input: &SendBounceRequest
    ) -> Result<SendBounceResponse, SendBounceError>;
    fn send_email(
        &self, 
        input: &SendEmailRequest
    ) -> Result<SendEmailResponse, SendEmailError>;
    fn send_raw_email(
        &self, 
        input: &SendRawEmailRequest
    ) -> Result<SendRawEmailResponse, SendRawEmailError>;
    fn set_active_receipt_rule_set(
        &self, 
        input: &SetActiveReceiptRuleSetRequest
    ) -> Result<SetActiveReceiptRuleSetResponse, SetActiveReceiptRuleSetError>;
    fn set_identity_dkim_enabled(
        &self, 
        input: &SetIdentityDkimEnabledRequest
    ) -> Result<SetIdentityDkimEnabledResponse, SetIdentityDkimEnabledError>;
    fn set_identity_feedback_forwarding_enabled(
        &self, 
        input: &SetIdentityFeedbackForwardingEnabledRequest
    ) -> Result<SetIdentityFeedbackForwardingEnabledResponse, SetIdentityFeedbackForwardingEnabledError>;
    fn set_identity_headers_in_notifications_enabled(
        &self, 
        input: &SetIdentityHeadersInNotificationsEnabledRequest
    ) -> Result<SetIdentityHeadersInNotificationsEnabledResponse, SetIdentityHeadersInNotificationsEnabledError>;
    fn set_identity_mail_from_domain(
        &self, 
        input: &SetIdentityMailFromDomainRequest
    ) -> Result<SetIdentityMailFromDomainResponse, SetIdentityMailFromDomainError>;
    fn set_identity_notification_topic(
        &self, 
        input: &SetIdentityNotificationTopicRequest
    ) -> Result<SetIdentityNotificationTopicResponse, SetIdentityNotificationTopicError>;
    fn set_receipt_rule_position(
        &self, 
        input: &SetReceiptRulePositionRequest
    ) -> Result<SetReceiptRulePositionResponse, SetReceiptRulePositionError>;
    fn update_configuration_set_event_destination(
        &self, 
        input: &UpdateConfigurationSetEventDestinationRequest
    ) -> Result<UpdateConfigurationSetEventDestinationResponse, UpdateConfigurationSetEventDestinationError>;
    fn update_receipt_rule(
        &self, 
        input: &UpdateReceiptRuleRequest
    ) -> Result<UpdateReceiptRuleResponse, UpdateReceiptRuleError>;
    fn verify_domain_dkim(
        &self, 
        input: &VerifyDomainDkimRequest
    ) -> Result<VerifyDomainDkimResponse, VerifyDomainDkimError>;
    fn verify_domain_identity(
        &self, 
        input: &VerifyDomainIdentityRequest
    ) -> Result<VerifyDomainIdentityResponse, VerifyDomainIdentityError>;
    fn verify_email_address(
        &self, 
        input: &VerifyEmailAddressRequest
    ) -> Result<(), VerifyEmailAddressError>;
    fn verify_email_identity(
        &self, 
        input: &VerifyEmailIdentityRequest
    ) -> Result<VerifyEmailIdentityResponse, VerifyEmailIdentityError>;
}

Trait representing the capabilities of the Amazon SES API. Amazon SES clients implement this trait.

Required Methods

fn clone_receipt_rule_set(
&self,
input: &CloneReceiptRuleSetRequest
) -> Result<CloneReceiptRuleSetResponse, CloneReceiptRuleSetError>

Creates a receipt rule set by cloning an existing one. All receipt rules and configurations are copied to the new receipt rule set and are completely independent of the source rule set.

For information about setting up rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn create_configuration_set(
&self,
input: &CreateConfigurationSetRequest
) -> Result<CreateConfigurationSetResponse, CreateConfigurationSetError>

Creates a configuration set.

Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn create_configuration_set_event_destination(
&self,
input: &CreateConfigurationSetEventDestinationRequest
) -> Result<CreateConfigurationSetEventDestinationResponse, CreateConfigurationSetEventDestinationError>

Creates a configuration set event destination.

When you create or update an event destination, you must provide one, and only one, destination. The destination can be either Amazon CloudWatch or Amazon Kinesis Firehose.

An event destination is the AWS service to which Amazon SES publishes the email sending events associated with a configuration set. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn create_receipt_filter(
&self,
input: &CreateReceiptFilterRequest
) -> Result<CreateReceiptFilterResponse, CreateReceiptFilterError>

Creates a new IP address filter.

For information about setting up IP address filters, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn create_receipt_rule(
&self,
input: &CreateReceiptRuleRequest
) -> Result<CreateReceiptRuleResponse, CreateReceiptRuleError>

Creates a receipt rule.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn create_receipt_rule_set(
&self,
input: &CreateReceiptRuleSetRequest
) -> Result<CreateReceiptRuleSetResponse, CreateReceiptRuleSetError>

Creates an empty receipt rule set.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_configuration_set(
&self,
input: &DeleteConfigurationSetRequest
) -> Result<DeleteConfigurationSetResponse, DeleteConfigurationSetError>

Deletes a configuration set.

Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_configuration_set_event_destination(
&self,
input: &DeleteConfigurationSetEventDestinationRequest
) -> Result<DeleteConfigurationSetEventDestinationResponse, DeleteConfigurationSetEventDestinationError>

Deletes a configuration set event destination.

Configuration set event destinations are associated with configuration sets, which enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_identity(
&self,
input: &DeleteIdentityRequest
) -> Result<DeleteIdentityResponse, DeleteIdentityError>

Deletes the specified identity (an email address or a domain) from the list of verified identities.

This action is throttled at one request per second.

fn delete_identity_policy(
&self,
input: &DeleteIdentityPolicyRequest
) -> Result<DeleteIdentityPolicyResponse, DeleteIdentityPolicyError>

Deletes the specified sending authorization policy for the given identity (an email address or a domain). This API returns successfully even if a policy with the specified name does not exist.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_receipt_filter(
&self,
input: &DeleteReceiptFilterRequest
) -> Result<DeleteReceiptFilterResponse, DeleteReceiptFilterError>

Deletes the specified IP address filter.

For information about managing IP address filters, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_receipt_rule(
&self,
input: &DeleteReceiptRuleRequest
) -> Result<DeleteReceiptRuleResponse, DeleteReceiptRuleError>

Deletes the specified receipt rule.

For information about managing receipt rules, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_receipt_rule_set(
&self,
input: &DeleteReceiptRuleSetRequest
) -> Result<DeleteReceiptRuleSetResponse, DeleteReceiptRuleSetError>

Deletes the specified receipt rule set and all of the receipt rules it contains.

The currently active rule set cannot be deleted.

For information about managing receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn delete_verified_email_address(
&self,
input: &DeleteVerifiedEmailAddressRequest
) -> Result<(), DeleteVerifiedEmailAddressError>

Deletes the specified email address from the list of verified addresses.

The DeleteVerifiedEmailAddress action is deprecated as of the May 15, 2012 release of Domain Verification. The DeleteIdentity action is now preferred.

This action is throttled at one request per second.

fn describe_active_receipt_rule_set(
&self,
input: &DescribeActiveReceiptRuleSetRequest
) -> Result<DescribeActiveReceiptRuleSetResponse, DescribeActiveReceiptRuleSetError>

Returns the metadata and receipt rules for the receipt rule set that is currently active.

For information about setting up receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn describe_configuration_set(
&self,
input: &DescribeConfigurationSetRequest
) -> Result<DescribeConfigurationSetResponse, DescribeConfigurationSetError>

Returns the details of the specified configuration set.

Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn describe_receipt_rule(
&self,
input: &DescribeReceiptRuleRequest
) -> Result<DescribeReceiptRuleResponse, DescribeReceiptRuleError>

Returns the details of the specified receipt rule.

For information about setting up receipt rules, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn describe_receipt_rule_set(
&self,
input: &DescribeReceiptRuleSetRequest
) -> Result<DescribeReceiptRuleSetResponse, DescribeReceiptRuleSetError>

Returns the details of the specified receipt rule set.

For information about managing receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn get_identity_dkim_attributes(
&self,
input: &GetIdentityDkimAttributesRequest
) -> Result<GetIdentityDkimAttributesResponse, GetIdentityDkimAttributesError>

Returns the current status of Easy DKIM signing for an entity. For domain name identities, this action also returns the DKIM tokens that are required for Easy DKIM signing, and whether Amazon SES has successfully verified that these tokens have been published.

This action takes a list of identities as input and returns the following information for each:

Whether Easy DKIM signing is enabled or disabled.

A set of DKIM tokens that represent the identity. If the identity is an email address, the tokens represent the domain of that address.

Whether Amazon SES has successfully verified the DKIM tokens published in the domain's DNS. This information is only returned for domain name identities, not for email addresses.

This action is throttled at one request per second and can only get DKIM attributes for up to 100 identities at a time.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

fn get_identity_mail_from_domain_attributes(
&self,
input: &GetIdentityMailFromDomainAttributesRequest
) -> Result<GetIdentityMailFromDomainAttributesResponse, GetIdentityMailFromDomainAttributesError>

Returns the custom MAIL FROM attributes for a list of identities (email addresses and/or domains).

This action is throttled at one request per second and can only get custom MAIL FROM attributes for up to 100 identities at a time.

fn get_identity_notification_attributes(
&self,
input: &GetIdentityNotificationAttributesRequest
) -> Result<GetIdentityNotificationAttributesResponse, GetIdentityNotificationAttributesError>

Given a list of verified identities (email addresses and/or domains), returns a structure describing identity notification attributes.

This action is throttled at one request per second and can only get notification attributes for up to 100 identities at a time.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

fn get_identity_policies(
&self,
input: &GetIdentityPoliciesRequest
) -> Result<GetIdentityPoliciesResponse, GetIdentityPoliciesError>

Returns the requested sending authorization policies for the given identity (an email address or a domain). The policies are returned as a map of policy names to policy contents. You can retrieve a maximum of 20 policies at a time.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn get_identity_verification_attributes(
&self,
input: &GetIdentityVerificationAttributesRequest
) -> Result<GetIdentityVerificationAttributesResponse, GetIdentityVerificationAttributesError>

Given a list of identities (email addresses and/or domains), returns the verification status and (for domain identities) the verification token for each identity.

This action is throttled at one request per second and can only get verification attributes for up to 100 identities at a time.

fn get_send_quota(&self) -> Result<GetSendQuotaResponse, GetSendQuotaError>

Returns the user's current sending limits.

This action is throttled at one request per second.

fn get_send_statistics(
&self
) -> Result<GetSendStatisticsResponse, GetSendStatisticsError>

Returns the user's sending statistics. The result is a list of data points, representing the last two weeks of sending activity.

Each data point in the list contains statistics for a 15-minute interval.

This action is throttled at one request per second.

fn list_configuration_sets(
&self,
input: &ListConfigurationSetsRequest
) -> Result<ListConfigurationSetsResponse, ListConfigurationSetsError>

Lists the configuration sets associated with your AWS account.

Configuration sets enable you to publish email sending events. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second and can return up to 50 configuration sets at a time.

fn list_identities(
&self,
input: &ListIdentitiesRequest
) -> Result<ListIdentitiesResponse, ListIdentitiesError>

Returns a list containing all of the identities (email addresses and domains) for your AWS account, regardless of verification status.

This action is throttled at one request per second.

fn list_identity_policies(
&self,
input: &ListIdentityPoliciesRequest
) -> Result<ListIdentityPoliciesResponse, ListIdentityPoliciesError>

Returns a list of sending authorization policies that are attached to the given identity (an email address or a domain). This API returns only a list. If you want the actual policy content, you can use GetIdentityPolicies.

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn list_receipt_filters(
&self,
input: &ListReceiptFiltersRequest
) -> Result<ListReceiptFiltersResponse, ListReceiptFiltersError>

Lists the IP address filters associated with your AWS account.

For information about managing IP address filters, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn list_receipt_rule_sets(
&self,
input: &ListReceiptRuleSetsRequest
) -> Result<ListReceiptRuleSetsResponse, ListReceiptRuleSetsError>

Lists the receipt rule sets that exist under your AWS account. If there are additional receipt rule sets to be retrieved, you will receive a NextToken that you can provide to the next call to ListReceiptRuleSets to retrieve the additional entries.

For information about managing receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn list_verified_email_addresses(
&self
) -> Result<ListVerifiedEmailAddressesResponse, ListVerifiedEmailAddressesError>

Returns a list containing all of the email addresses that have been verified.

The ListVerifiedEmailAddresses action is deprecated as of the May 15, 2012 release of Domain Verification. The ListIdentities action is now preferred.

This action is throttled at one request per second.

fn put_identity_policy(
&self,
input: &PutIdentityPolicyRequest
) -> Result<PutIdentityPolicyResponse, PutIdentityPolicyError>

Adds or updates a sending authorization policy for the specified identity (an email address or a domain).

This API is for the identity owner only. If you have not verified the identity, this API will return an error.

Sending authorization is a feature that enables an identity owner to authorize other senders to use its identities. For information about using sending authorization, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn reorder_receipt_rule_set(
&self,
input: &ReorderReceiptRuleSetRequest
) -> Result<ReorderReceiptRuleSetResponse, ReorderReceiptRuleSetError>

Reorders the receipt rules within a receipt rule set.

All of the rules in the rule set must be represented in this request. That is, this API will return an error if the reorder request doesn't explicitly position all of the rules.

For information about managing receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn send_bounce(
&self,
input: &SendBounceRequest
) -> Result<SendBounceResponse, SendBounceError>

Generates and sends a bounce message to the sender of an email you received through Amazon SES. You can only use this API on an email up to 24 hours after you receive it.

You cannot use this API to send generic bounces for mail that was not received by Amazon SES.

For information about receiving email through Amazon SES, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn send_email(
&self,
input: &SendEmailRequest
) -> Result<SendEmailResponse, SendEmailError>

Composes an email message based on input data, and then immediately queues the message for sending.

There are several important points to know about SendEmail:

You can only send email from verified email addresses and domains; otherwise, you will get an "Email address not verified" error. If your account is still in the Amazon SES sandbox, you must also verify every recipient email address except for the recipients provided by the Amazon SES mailbox simulator. For more information, go to the Amazon SES Developer Guide.

The total size of the message cannot exceed 10 MB. This includes any attachments that are part of the message.

Amazon SES has a limit on the total number of recipients per message. The combined number of To:, CC: and BCC: email addresses cannot exceed 50. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call Amazon SES repeatedly to send the message to each group.

For every message that you send, the total number of recipients (To:, CC: and BCC:) is counted against your sending quota - the maximum number of emails you can send in a 24-hour period. For information about your sending quota, go to the Amazon SES Developer Guide.

fn send_raw_email(
&self,
input: &SendRawEmailRequest
) -> Result<SendRawEmailResponse, SendRawEmailError>

Sends an email message, with header and content specified by the client. The SendRawEmail action is useful for sending multipart MIME emails. The raw text of the message must comply with Internet email standards; otherwise, the message cannot be sent.

There are several important points to know about SendRawEmail:

You can only send email from verified email addresses and domains; otherwise, you will get an "Email address not verified" error. If your account is still in the Amazon SES sandbox, you must also verify every recipient email address except for the recipients provided by the Amazon SES mailbox simulator. For more information, go to the Amazon SES Developer Guide.

The total size of the message cannot exceed 10 MB. This includes any attachments that are part of the message.

Amazon SES has a limit on the total number of recipients per message. The combined number of To:, CC: and BCC: email addresses cannot exceed 50. If you need to send an email message to a larger audience, you can divide your recipient list into groups of 50 or fewer, and then call Amazon SES repeatedly to send the message to each group.

The To:, CC:, and BCC: headers in the raw message can contain a group list. Note that each recipient in a group list counts towards the 50-recipient limit.

Amazon SES overrides any Message-ID and Date headers you provide.

For every message that you send, the total number of recipients (To:, CC: and BCC:) is counted against your sending quota - the maximum number of emails you can send in a 24-hour period. For information about your sending quota, go to the Amazon SES Developer Guide.

If you are using sending authorization to send on behalf of another user, SendRawEmail enables you to specify the cross-account identity for the email's "Source," "From," and "Return-Path" parameters in one of two ways: you can pass optional parameters SourceArn, FromArn, and/or ReturnPathArn to the API, or you can include the following X-headers in the header of your raw email:

X-SES-SOURCE-ARN

X-SES-FROM-ARN

X-SES-RETURN-PATH-ARN

Do not include these X-headers in the DKIM signature, because they are removed by Amazon SES before sending the email.

For the most common sending authorization use case, we recommend that you specify the SourceIdentityArn and do not specify either the FromIdentityArn or ReturnPathIdentityArn. (The same note applies to the corresponding X-headers.) If you only specify the SourceIdentityArn, Amazon SES will simply set the "From" address and the "Return Path" address to the identity specified in SourceIdentityArn. For more information about sending authorization, see the Amazon SES Developer Guide.

fn set_active_receipt_rule_set(
&self,
input: &SetActiveReceiptRuleSetRequest
) -> Result<SetActiveReceiptRuleSetResponse, SetActiveReceiptRuleSetError>

Sets the specified receipt rule set as the active receipt rule set.

To disable your email-receiving through Amazon SES completely, you can call this API with RuleSetName set to null.

For information about managing receipt rule sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn set_identity_dkim_enabled(
&self,
input: &SetIdentityDkimEnabledRequest
) -> Result<SetIdentityDkimEnabledResponse, SetIdentityDkimEnabledError>

Enables or disables Easy DKIM signing of email sent from an identity:

If Easy DKIM signing is enabled for a domain name identity (e.g., example.com), then Amazon SES will DKIM-sign all email sent by addresses under that domain name (e.g., user@example.com).

If Easy DKIM signing is enabled for an email address, then Amazon SES will DKIM-sign all email sent by that email address.

For email addresses (e.g., user@example.com), you can only enable Easy DKIM signing if the corresponding domain (e.g., example.com) has been set up for Easy DKIM using the AWS Console or the VerifyDomainDkim action.

This action is throttled at one request per second.

For more information about Easy DKIM signing, go to the Amazon SES Developer Guide.

fn set_identity_feedback_forwarding_enabled(
&self,
input: &SetIdentityFeedbackForwardingEnabledRequest
) -> Result<SetIdentityFeedbackForwardingEnabledResponse, SetIdentityFeedbackForwardingEnabledError>

Given an identity (an email address or a domain), enables or disables whether Amazon SES forwards bounce and complaint notifications as email. Feedback forwarding can only be disabled when Amazon Simple Notification Service (Amazon SNS) topics are specified for both bounces and complaints.

Feedback forwarding does not apply to delivery notifications. Delivery notifications are only available through Amazon SNS.

This action is throttled at one request per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

fn set_identity_headers_in_notifications_enabled(
&self,
input: &SetIdentityHeadersInNotificationsEnabledRequest
) -> Result<SetIdentityHeadersInNotificationsEnabledResponse, SetIdentityHeadersInNotificationsEnabledError>

Given an identity (an email address or a domain), sets whether Amazon SES includes the original email headers in the Amazon Simple Notification Service (Amazon SNS) notifications of a specified type.

This action is throttled at one request per second.

For more information about using notifications with Amazon SES, see the Amazon SES Developer Guide.

fn set_identity_mail_from_domain(
&self,
input: &SetIdentityMailFromDomainRequest
) -> Result<SetIdentityMailFromDomainResponse, SetIdentityMailFromDomainError>

Enables or disables the custom MAIL FROM domain setup for a verified identity (an email address or a domain).

To send emails using the specified MAIL FROM domain, you must add an MX record to your MAIL FROM domain's DNS settings. If you want your emails to pass Sender Policy Framework (SPF) checks, you must also add or update an SPF record. For more information, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn set_identity_notification_topic(
&self,
input: &SetIdentityNotificationTopicRequest
) -> Result<SetIdentityNotificationTopicResponse, SetIdentityNotificationTopicError>

Given an identity (an email address or a domain), sets the Amazon Simple Notification Service (Amazon SNS) topic to which Amazon SES will publish bounce, complaint, and/or delivery notifications for emails sent with that identity as the Source.

Unless feedback forwarding is enabled, you must specify Amazon SNS topics for bounce and complaint notifications. For more information, see SetIdentityFeedbackForwardingEnabled.

This action is throttled at one request per second.

For more information about feedback notification, see the Amazon SES Developer Guide.

fn set_receipt_rule_position(
&self,
input: &SetReceiptRulePositionRequest
) -> Result<SetReceiptRulePositionResponse, SetReceiptRulePositionError>

Sets the position of the specified receipt rule in the receipt rule set.

For information about managing receipt rules, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn update_configuration_set_event_destination(
&self,
input: &UpdateConfigurationSetEventDestinationRequest
) -> Result<UpdateConfigurationSetEventDestinationResponse, UpdateConfigurationSetEventDestinationError>

Updates the event destination of a configuration set.

When you create or update an event destination, you must provide one, and only one, destination. The destination can be either Amazon CloudWatch or Amazon Kinesis Firehose.

Event destinations are associated with configuration sets, which enable you to publish email sending events to Amazon CloudWatch or Amazon Kinesis Firehose. For information about using configuration sets, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn update_receipt_rule(
&self,
input: &UpdateReceiptRuleRequest
) -> Result<UpdateReceiptRuleResponse, UpdateReceiptRuleError>

Updates a receipt rule.

For information about managing receipt rules, see the Amazon SES Developer Guide.

This action is throttled at one request per second.

fn verify_domain_dkim(
&self,
input: &VerifyDomainDkimRequest
) -> Result<VerifyDomainDkimResponse, VerifyDomainDkimError>

Returns a set of DKIM tokens for a domain. DKIM tokens are character strings that represent your domain's identity. Using these tokens, you will need to create DNS CNAME records that point to DKIM public keys hosted by Amazon SES. Amazon Web Services will eventually detect that you have updated your DNS records; this detection process may take up to 72 hours. Upon successful detection, Amazon SES will be able to DKIM-sign email originating from that domain.

This action is throttled at one request per second.

To enable or disable Easy DKIM signing for a domain, use the SetIdentityDkimEnabled action.

For more information about creating DNS records using DKIM tokens, go to the Amazon SES Developer Guide.

fn verify_domain_identity(
&self,
input: &VerifyDomainIdentityRequest
) -> Result<VerifyDomainIdentityResponse, VerifyDomainIdentityError>

Verifies a domain.

This action is throttled at one request per second.

fn verify_email_address(
&self,
input: &VerifyEmailAddressRequest
) -> Result<(), VerifyEmailAddressError>

Verifies an email address. This action causes a confirmation email message to be sent to the specified address.

The VerifyEmailAddress action is deprecated as of the May 15, 2012 release of Domain Verification. The VerifyEmailIdentity action is now preferred.

This action is throttled at one request per second.

fn verify_email_identity(
&self,
input: &VerifyEmailIdentityRequest
) -> Result<VerifyEmailIdentityResponse, VerifyEmailIdentityError>

Verifies an email address. This action causes a confirmation email message to be sent to the specified address.

This action is throttled at one request per second.