Struct rusoto_ses::SesClient

pub struct SesClient { /* fields omitted */ }

A client for the Amazon SES API.

Methods

impl SesClient

pub fn new(region: Region) -> SesClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SesClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,